requirement.txt
python-telegram-bot==20.6
openai==1.51.0
python-dotenv
